User/derel/data validation

+++ api/Controllers/StockController.cs
@@ -1,8 +1,10 @@
 using api.Data;
 using api.Dtos.Stock;
+using api.Helpers;
 using api.Interfaces;
 using api.Mappers;
 using api.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -21,21 +23,32 @@ namespace api.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll()
+        [Authorize]
+        public async Task<IActionResult> GetAll([FromQuery] QueryObject query)
         {
+            
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            
             //var stocks = await _context.Stocks.ToListAsync();
 
             //Using Repository
-            var stocks = await _stockRepository.GetAllAsync();
+            var stocks = await _stockRepository.GetAllAsync(query);
 
             var stockDtos = stocks.Select(s => s.ToStockDto());
 
             return Ok(stockDtos);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             //var stock = await _context.Stocks.FindAsync(id);
 
             var stock = await _stockRepository.GetByIdAsync(id);
@@ -49,8 +62,14 @@ namespace api.Controllers
         }
 
         [HttpPost]
+        [Authorize]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var createstock = stockDto.ToStockFromCreateDto();
 
             await _stockRepository.CreateAsync(createstock);
@@ -59,10 +78,16 @@ namespace api.Controllers
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updatestockDto)
         {
-            var stockModel = await _stockRepository.UpdateAsync(id, updatestockDto);
+            
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            
+            var stockModel = await _stockRepository.UpdateAsync(id, updatestockDto.ToStockFromUpdateDto());
 
             if (stockModel == null)
             {
@@ -73,11 +98,15 @@ namespace api.Controllers
         }
 
         [HttpDelete]
-        [Route("{id}")]
-
-
-        public async Task<IActionResult> Delete(int id)
+        [Route("{id:int}")]
+        public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+            
             var deleted = await _stockRepository.DeleteAsync(id);
 
             if (!deleted)
